requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
bandit==1.8.6
bidict==0.23.1
blinker==1.9.0
boolean.py==5.0
Brotli==1.1.0
CacheControl==0.14.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
ConfigArgParse==1.7.1
cyclonedx-python-lib==9.1.0
defusedxml==0.7.1
distro==1.9.0
execnet==2.1.1
Faker==37.6.0
fastapi==0.116.1
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
gevent==25.5.1
geventhttpclient==2.3.4
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
license-expression==30.4.4
locust==2.39.1
locust-cloud==1.26.3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.1
multidict==6.6.4
openai==1.104.2
packageurl-python==0.17.5
packaging==25.0
pact-python==2.3.3
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.4.0
pluggy==1.6.0
propcache==0.3.2
psutil==7.0.0
py-cpuinfo==9.0.0
py-serializable==2.1.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
pytest-benchmark==5.1.0
pytest-xdist==3.8.0
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
PyYAML==6.0.2
pyzmq==27.0.2
referencing==0.36.2
requests==2.32.5
responses==0.25.8
rich==14.1.0
rpds-py==0.27.1
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.47.3
stevedore==5.5.0
toml==0.10.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.1
zope.event==5.1.1
zope.interface==7.2
